Changes: - Moved invalid description validation from commands to Task - Added tests for many more classes - Moved common task test constants to new interface file
Fixes:
- order of task operations preventing size miscalculation if misused

+++ src/test/java/duke/DukeUtilTest.java
@@ -35,7 +35,7 @@ class DukeUtilTest {
                 .range(2, 30)
                 .filter(x -> IntStream
                     .range(2, x)
-                    .noneMatch(y -> (double) x / y ==  x / y))
+                    .noneMatch(y -> (double) x / (double) y ==  x / y))
                 .mapToObj(primeNum -> primeNum + "")
                 .toArray(size -> new String[15]);
